+++ ntex/src/http/service.rs
@@ -1,8 +1,8 @@
-use std::{cell::Cell, cell::RefCell, error, fmt, marker, rc::Rc};
+use std::{cell::Cell, cell::RefCell, error, fmt, marker, rc::Rc, task::Context};
 
 use crate::io::{types, Filter, Io, IoRef};
 use crate::service::{IntoServiceFactory, Service, ServiceCtx, ServiceFactory};
-use crate::{channel::oneshot, util::join, util::select, util::HashSet};
+use crate::{channel::oneshot, util::join, util::HashSet};
 
 use super::body::MessageBody;
 use super::builder::HttpServiceBuilder;
@@ -312,12 +312,16 @@ where
     }
 
     #[inline]
-    async fn not_ready(&self) {
+    fn poll(&self, cx: &mut Context<'_>) -> Result<(), Self::Error> {
         let cfg = self.config.as_ref();
-        select(cfg.control.not_ready(), cfg.service.not_ready()).await;
+        cfg.control
+            .poll(cx)
+            .map_err(|e| DispatchError::Control(Box::new(e)))?;
+        cfg.service
+            .poll(cx)
+            .map_err(|e| DispatchError::Service(Box::new(e)))
     }
 
-    #[inline]
     async fn shutdown(&self) {
         self.config.shutdown();