+++ ntex/src/web/middleware/defaultheaders.rs
@@ -110,6 +110,7 @@ where
     type Response = WebResponse;
     type Error = S::Error;
 
+    crate::forward_poll!(service);
     crate::forward_ready!(service);
     crate::forward_shutdown!(service);