+++ ntex/src/web/middleware/logger.rs
@@ -139,6 +139,7 @@ where
     type Response = WebResponse;
     type Error = S::Error;
 
+    crate::forward_poll!(service);
     crate::forward_ready!(service);
     crate::forward_shutdown!(service);